fix: serve schemas at canonical top-level URLs (no /schemas/ prefix)
Fixes #234 — the site redesign broke all legacy schema URLs by adding
a /schemas/ prefix. XML parsers using xs:import/xs:include got 404s
on the old canonical paths like /19139/-/gco/1.0/gco.xsd.

Changes:
- BuildSourceGenerator plugin registers schema data files from the
  schemas/ submodule as Jekyll static files at canonical top-level paths
- Redirect generator copies actual data files to legacy paths (XML
  parsers can't follow HTML redirects)
- Shared HtmlHelper module in models.rb eliminates duplicated esc()
  and render_resources() across 3 page classes
- Removed /schemas/ prefix from all links (redirects, resources JS,
  index, docs, about pages)
- Streamlined Makefile (removed build-all, distclean, merged update
  targets)
- Removed manual cp steps from CI — everything handled by Jekyll plugins
