+++ Makefile
@@ -1,29 +1,22 @@
 SHELL := /bin/bash
 LUTAML_BUNDLE := BUNDLE_GEMFILE=schemas/Gemfile BUNDLE_PATH=vendor/bundle-lutaml
-SITE_BUNDLE := BUNDLE_GEMFILE=Gemfile BUNDLE_PATH=vendor/bundle
 
 all: configs lxr-spas _site
 
-# Generate per-schema config files, index JSON, and Makefile.spa
-schemas_index.json: generate_configs.rb
+# Generate per-schema config files and index JSON
+configs:
 	ruby generate_configs.rb
-
-configs: schemas_index.json ;
 
 # Build all LXR + SPA packages
 lxr-spas: configs
 	@GEM_DIR=$$($(LUTAML_BUNDLE) bundle show lutaml-jsonschema 2>/dev/null) && \
 		[ -d "$$GEM_DIR/frontend/dist" ] || { cd "$$GEM_DIR/frontend" && npm install && npm run build; } || true
-	$(MAKE) build-all
-
-build-all:
 	$(MAKE) -j4 $(LXR_FILES) $(SPA_FILES)
 
 CONFIGS   := $(wildcard configs/*.yml)
 LXR_FILES := $(patsubst configs/%.yml,build/%.lxr,$(CONFIGS))
 SPA_FILES :=
 
-# Build a single LXR package from a config
 build/%.lxr: configs/%.yml
 	mkdir -p build
 	$(LUTAML_BUNDLE) bundle exec lutaml-xsd build from-config $< \
@@ -32,26 +25,19 @@ build/%.lxr: configs/%.yml
 # SPA build rules (auto-generated by generate_configs.rb)
 -include configs/Makefile.spa
 
-# Build Jekyll site
+# Build Jekyll site (SPA pages + schema data registered by BuildSourceGenerator plugin)
 _site: lxr-spas
 	JEKYLL_ENV=production bundle exec jekyll build
-	cp -r site/* _site/ 2>/dev/null || true
 
-# Dev server
 serve:
-	$(SITE_BUNDLE) bundle exec jekyll serve
+	bundle exec jekyll serve
 
 clean:
 	rm -rf build site _site configs schemas_index.json resources_index.json
 
-distclean: clean
-	rm -rf schemas/_site vendor/bundle-lutaml
-
-update-init:
+update:
 	git submodule update --init
-
-update-modules:
 	git submodule foreach git checkout main
 	git submodule foreach git pull origin main
 
-.PHONY: all configs build-all lxr-spas clean distclean serve update-init update-modules
+.PHONY: all configs lxr-spas clean serve update
