+++ _pages/about.html
@@ -89,7 +89,7 @@ permalink: /about/
         </li>
         <li>
           <strong>JSON Schemas</strong> — served under
-          <code>/schemas/json/{standard}/…</code>. Use the <code>$id</code>
+          <code>/json/{standard}/…</code>. Use the <code>$id</code>
           URI for <code>$ref</code> resolution.
         </li>
         <li>
