+++ _pages/docs.html
@@ -37,7 +37,7 @@ permalink: /docs/
         <span class="doc-card__link">Browse resources</span>
       </a>
 
-      <a href="/schemas/" class="doc-card">
+      <a href="https://github.com/ISO-TC211/schemas" class="doc-card">
         <div class="doc-card__icon">
           <svg fill="none" stroke="currentColor" viewBox="0 0 24 24" stroke-width="1.5"><path stroke-linecap="round" stroke-linejoin="round" d="M3.75 9.776c.112-.017.227-.026.344-.026h15.812c.117 0 .232.009.344.026m-16.5 0a2.25 2.25 0 00-1.883 2.542l.857 6a2.25 2.25 0 002.227 1.932H19.05a2.25 2.25 0 002.227-1.932l.857-6a2.25 2.25 0 00-1.883-2.542m-16.5 0V6A2.25 2.25 0 016 3.75h3.879a1.5 1.5 0 011.06.44l2.122 2.12a1.5 1.5 0 001.06.44H18A2.25 2.25 0 0120.25 9v.776"/></svg>
         </div>
@@ -64,7 +64,7 @@ permalink: /docs/
         </div>
         <div class="url-box__item--dark url-box__item">
           <p class="url-box__label">JSON schema location</p>
-          <code>https://schemas.isotc211.org/schemas/json/<span class="dim">{Standard}</span>/<span class="dim">{Part}</span>/<span class="dim">{Module}</span>/<span class="dim">{Version}</span>/<span class="dim">{Schema}</span>.json</code>
+          <code>https://schemas.isotc211.org/json/<span class="dim">{Standard}</span>/<span class="dim">{Part}</span>/<span class="dim">{Module}</span>/<span class="dim">{Version}</span>/<span class="dim">{Schema}</span>.json</code>
         </div>
       </div>
     </div>
