+++ _pages/index.html
@@ -44,9 +44,9 @@ permalink: /
     <h2 class="resources-section__title">Resources</h2>
     <ul class="resources-list">
       <li>
-        <a href="/schemas/">
+        <a href="https://github.com/ISO-TC211/schemas">
           <svg fill="none" stroke="currentColor" viewBox="0 0 24 24" stroke-width="2"><path stroke-linecap="round" stroke-linejoin="round" d="M3 7v10a2 2 0 002 2h14a2 2 0 002-2V9a2 2 0 00-2-2h-6l-2-2H5a2 2 0 00-2 2z"/></svg>
-          Browse raw XSD files
+          Browse raw files on GitHub
         </a>
       </li>
       <li>
